pyproject.toml
[tool.poetry]
name = "nibiru"
version = "0.4.2"
description = "Python SDK for interacting with Nibiru."
authors = ["Nibiru Chain <dev@nibiru.fi>"]
license = "MIT"
maintainers = [
  "NibiruHeisenberg <dev@nibiru.fi>",
  "matthiasmat <dev@nibiru.fi>",
  "onikonychev <dev@nibiru.fi>",
  "Unique-Divine <dev@nibiru.fi>"
]
readme = "README.md" # similar to long_description in setuptools
repository = "https://github.com/NibiruChain/sdk-py"
keywords = ["nibiru", "blockchain", "sdk", "python", "cosmos"]
classifiers = [
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Intended Audience :: Developers"
]
packages = [
    { include = "nibiru" },
]

[tool.poetry.dependencies]
python = "^3.9"
grpcio = "^1.47.0"
asyncio = "^3.4.3"
bech32 = "^1.2.0"
python-dotenv = "^0.20.0"
protobuf = "3.20.1" # see README
requests = "^2.28.1"
aiohttp = "^3.8.1"
ecdsa = "^0.18.0"
mnemonic = "^0.20"
hdwallets = "^0.1.2"
bip32 = "^3.1"
eip712-structs = "^1.1.0"
coincurve = "^17.0.0"
aiocron = "^1.8"
pytest = "^7.1.2"
pre-commit = "^2.20.0"
nibiru-proto = "^0.14.1"
shutup = "^0.2.0"
websocket-client = "^1.4.0"
coverage = "^6.4.4"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
flake8 = "^5.0.4"
importlib-metadata = {version = "^1.0", python = "<3.8"}


[tool.black]
line-length = 88
skip-string-normalization = true
include = '.*\.pyi?$'
target-version = ['py39']
extend-exclude = '''
^/(
    \.vscode
  | nibiru/proto
)$
'''

[tool.isort]
profile = "black"
py_version = 39
skip_glob = ["nibiru/proto/*"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
